Minimal patch to allow the trie logic to handle EXACTFU nodes
The old logic was that we would trie EXACTF nodes only when they were
utf8. When the /u support was added things were changed so that EXACTFU
nodes are produced in this case, however they are also produced when
we are using /u on a non UTF pattern. This patch teaches the trie logic
to handle this case, so it is also now converted to a trie.

Note, this patch is deliberately minimal, we should in a future patch
review this logic for further improvements.